requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0

# Encryption
pycryptodome>=3.15.0
cryptography>=38.0.0

# Embeddings & ML
sentence-transformers>=2.2.0
torch>=1.13.0
transformers>=4.25.0

# API & Web
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5

# Utilities
python-dotenv>=0.19.0
requests>=2.28.0
tqdm>=4.64.0
